requirements-dev.txt
# Development dependencies
-r requirements.txt

pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-httpx>=0.30.0
mypy>=1.8.0
ruff>=0.3.0
black>=24.0.0
